==> Data_Cleaning/Step_1_Data_Processing.ipynb <==
#Import Libraries

import os
import pandas as pd
import seaborn as sns
import matplotlib.pylab as plt
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

#Import Bike Data

df_bike = pd.DataFrame()
path = '/content/drive/MyDrive/CIV1498 Final Project/Data/data/ridership/release'
file_lst = os.listdir(path)
file_2017 = []
file_2018 = []
file_2019 = []
file_2020 = []
file_2021 = []
for file_name in file_lst:
  if '2017' in file_name:
    file_2017.append(file_name)
  if '2018' in file_name:
    file_2018.append(file_name)
  if '2019' in file_name:
    file_2019.append(file_name)
  if '2020' in file_name:
    file_2020.append(file_name)
  if '2021' in file_name:
    file_2021.append(file_name)

#what we want
header = ['trip_id', 'user_type', 'start_time', 'end_time', 'duration', 'start_station_id', 'start_station_name', 'end_station_id', 'end_station_name']

#2017 data
df_bike_2017 = pd.DataFrame()
for file_name in file_2017:
  df = pd.read_csv(path+'/'+file_name)
  df_bike_2017 = pd.concat([df_bike_2017, df], axis = 0)

df_bike_2017.columns = ['trip_id', 'start_time', 'end_time', 'duration', 'start_station_id', 'start_station_name', 'end_station_id', 'end_station_name', 'user_type']
df_bike_2017 = df_bike_2017[header]
df_bike_2017.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/bike_2017.csv', index = False)

#2018 data
df_bike_2018 = pd.DataFrame()
for file_name in file_2018:
  df = pd.read_csv(path+'/'+file_name)
  df_bike_2018 = pd.concat([df_bike_2018, df], axis = 0)

df_bike_2018.columns = ['trip_id', 'duration', 'start_station_id', 'start_time', 'start_station_name', 'end_time', 'end_station_id', 'end_station_name', 'user_type']
df_bike_2018 = df_bike_2018[header]
df_bike_2018.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/bike_2018.csv', index = False)

#2019 data
df_bike_2019 = pd.DataFrame()
for file_name in file_2019:
  df = pd.read_csv(path+'/'+file_name)
  df_bike_2019 = pd.concat([df_bike_2019, df], axis = 0)

df_bike_2019.columns = ['trip_id', '__', 'duration', 'start_station_id', 'start_time', 'start_station_name', 'end_station_id', 'end_time', 'end_station_name', '__', 'user_type']
df_bike_2019 = df_bike_2019[header]
df_bike_2019.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/bike_2019.csv', index = False)

#2020 data
df_bike_2020 = pd.DataFrame()
for file_name in file_2020:
  df = pd.read_csv(path+'/'+file_name)
  df_bike_2020 = pd.concat([df_bike_2020, df], axis = 0)

df_bike_2020.columns = ['trip_id', '__', 'duration', 'start_station_id', 'start_time', 'start_station_name', 'end_station_id', 'end_time', 'end_station_name', '__', 'user_type']
df_bike_2020 = df_bike_2020[header]
df_bike_2020.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/bike_2020.csv', index = False)

#2021 data
df_bike_2021 = pd.DataFrame()
for file_name in file_2021:
  df = pd.read_csv(path+'/'+file_name)
  df_bike_2021 = pd.concat([df_bike_2021, df], axis = 0)

df_bike_2021['trip_id'] = df_bike_2021['ï»¿Trip Id'].fillna(0)+df_bike_2021['Trip Id'].fillna(0)
df_bike_2021.columns = ['__', 'duration', 'start_station_id', 'start_time', 'start_station_name', 'end_station_id', 'end_time', 'end_station_name', '__', 'user_type', '__', 'trip_id']
df_bike_2021 = df_bike_2021[header]
df_bike_2021.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/bike_2021.csv', index = False)

#Import Weather Data

df_weather = pd.DataFrame()
path = '/content/drive/MyDrive/CIV1498 Final Project/Data/data/weather'
file_lst = os.listdir(path)

file_2017 = []
file_2018 = []
file_2019 = []
file_2020 = []
file_2021 = []
for file_name in file_lst:
  if '2017' in file_name:
    file_2017.append(file_name)
  if '2018' in file_name:
    file_2018.append(file_name)
  if '2019' in file_name:
    file_2019.append(file_name)
  if '2020' in file_name:
    file_2020.append(file_name)
  if '2021' in file_name:
    file_2021.append(file_name)

#what we want to keep
header = ['Longitude (x)', 
          'Latitude (y)', 
          'Station Name',
          'Date/Time', 
          'Year', 
          'Month', 
          'Day', 
          'Time', 
          'Temp (°C)',
          'Dew Point Temp (°C)', 
          'Rel Hum (%)',
          'Wind Dir (10s deg)',
          'Wind Spd (km/h)',
          'Visibility (km)',
          'Stn Press (kPa)',
          'Weather']

#2017 data
df_weather_2017 = pd.DataFrame()
for file_name in file_2017:
  df = pd.read_csv(path+'/'+file_name)
  df_weather_2017 = pd.concat([df_weather_2017, df], axis = 0)

df_weather_2017 = df_weather_2017[header]
df_weather_2017.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/weather_2017.csv', index = False)

#2018 data
df_weather_2018 = pd.DataFrame()
for file_name in file_2018:
  df = pd.read_csv(path+'/'+file_name)
  df_weather_2018 = pd.concat([df_weather_2018, df], axis = 0)

df_weather_2018 = df_weather_2018[header]
df_weather_2018.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/weather_2018.csv', index = False)

#2019 data
df_weather_2019 = pd.DataFrame()
for file_name in file_2019:
  df = pd.read_csv(path+'/'+file_name)
  df_weather_2019 = pd.concat([df_weather_2019, df], axis = 0)

df_weather_2019 = df_weather_2019[header]
df_weather_2019.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/weather_2019.csv', index = False)

#2020 data
df_weather_2020 = pd.DataFrame()
for file_name in file_2020:
  df = pd.read_csv(path+'/'+file_name)
  df_weather_2020 = pd.concat([df_weather_2020, df], axis = 0)

df_weather_2020 = df_weather_2020[header]
df_weather_2020.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/weather_2020.csv', index = False)

header = ['Longitude (x)', 
          'Latitude (y)', 
          'Station Name',
          'Date/Time (LST)', 
          'Year', 
          'Month', 
          'Day', 
          'Time (LST)', 
          'Temp (°C)',
          'Dew Point Temp (°C)', 
          'Rel Hum (%)',
          'Wind Dir (10s deg)',
          'Wind Spd (km/h)',
          'Visibility (km)',
          'Stn Press (kPa)',
          'Weather']

#2021 data
df_weather_2021 = pd.DataFrame()
for file_name in file_2021:
  df = pd.read_csv(path+'/'+file_name)
  df_weather_2021 = pd.concat([df_weather_2021, df], axis = 0)

df_weather_2021 = df_weather_2021[header]
df_weather_2021.columns = ['Longitude (x)', 
          'Latitude (y)', 
          'Station Name',
          'Date/Time', 
          'Year', 
          'Month', 
          'Day', 
          'Time', 
          'Temp (°C)',
          'Dew Point Temp (°C)', 
          'Rel Hum (%)',
          'Wind Dir (10s deg)',
          'Wind Spd (km/h)',
          'Visibility (km)',
          'Stn Press (kPa)',
          'Weather']
df_weather_2021.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/weather_2021.csv', index = False)

#Import Station Data

path = '/content/drive/MyDrive/CIV1498 Final Project/Data/data/stations/bikeshare_stations.csv'
df_station = pd.read_csv(path)
df_station.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data/stations.csv')
