==> Data_Cleaning/Step_3_weather_and_trip.ipynb <==
#Import Libraries

import os
import pandas as pd
import seaborn as sns
import matplotlib.pylab as plt
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

#Import Data

df_bike = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/trip_df.csv')

df_bike_copy = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/trip_df.csv')

df_bike['start_time'] = pd.to_datetime(df_bike['start_time'])

df_weather = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/weather_df.csv')
df_weather_copy = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/weather_df.csv')

df_weather['Unnamed: 0'] = pd.to_datetime(df_weather['Unnamed: 0'])
df_weather = df_weather.sort_values(by = 'Unnamed: 0').reset_index(inplace = False)
df_weather['Time'] = df_weather['Unnamed: 0']
df_weather = df_weather.drop(columns = ['Unnamed: 0'])
df_weather.head()

df_bike_weather = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/daily_rides_and_weather.csv')
df_bike_weather_copy = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data2/daily_rides_and_weather.csv')

df_bike_weather_copy

#weather data

hour_of_day = []
for time in df_weather['Time']:
  hour_of_day.append(time.hour + time.minute/60 + time.second/60**2)
df_weather['hour_of_day'] = hour_of_day

def period(x):
  bins = np.array([0,6,10,15,19,24])  
  pod = np.digitize(x,bins)
  return pod

df_weather['period'] = df_weather['hour_of_day'].apply(period)

df_weather['Weather'] = df_weather['Weather'].fillna('Clear')
df_weather['Weather'].unique()

df_weather['Weather'] = df_weather['Weather'].replace('Haze', 'Fog')
df_weather['Weather'] = df_weather['Weather'].replace(['Rain,Fog', 'Rain', 'Snow', 'Moderate Rain',
       'Moderate Rain,Fog', 'Moderate Snow', 'Heavy Rain,Fog',
       'Thunderstorms,Rain,Fog', 'Thunderstorms,Rain', 'Thunderstorms',
       'Thunderstorms,Moderate Rain', 'Thunderstorms,Moderate Rain,Fog',
       'Thunderstorms,Heavy Rain,Fog', 'Rain,Snow', 'Freezing Rain,Fog',
       'Freezing Rain,Snow', 'Freezing Rain', 'Snow,Blowing Snow',
       'Heavy Snow', 'Haze,Blowing Snow', 'Thunderstorms,Fog',
       'Thunderstorms,Heavy Rain'], 'Precipitation')
df_weather['Weather'].unique()

df_weather.isnull().sum()

month_weather = df_weather.groupby('Month').mean()[['Temp (°C)', 'Dew Point Temp (°C)', 'Rel Hum (%)', 'Wind Dir (10s deg)', 'Wind Spd (km/h)', 'Visibility (km)', 'Stn Press (kPa)']]

month_weather = month_weather.reset_index()

month_weather

(15.495685+13.355135)/2


month_weather['Wind Dir (10s deg)'].mean()

df_weather

weather = []
temp = []
dew_temp = []
rh = []
wind_dir = []
wind_speed = []
visibility = []
pressure = []

for index, group in df_weather.groupby(['Year', 'Month', 'Day', 'period']):
  month = int(group['Month'].mean())
  weather.append(group['Weather'].mode()[0])

  #temperature
  if group['Temp (°C)'].mean() == np.nan:
    temp.append(month_weather[month_weather['Month'] == month]['Temp (°C)'])
  else:
    temp.append(group['Temp (°C)'].mean())

  #dew point temp
  if group['Dew Point Temp (°C)'].mean() == np.nan:
    dew_temp.append(month_weather[month_weather['Month'] == month]['Dew Point Temp (°C)'])
  else:
    dew_temp.append(group['Dew Point Temp (°C)'].mean())

  #humidity
  if group['Rel Hum (%)'].mean() == np.nan:
    rh.append(month_weather[month_weather['Month'] == month]['Rel Hum (%)'])
  else:
    rh.append(group['Rel Hum (%)'].mean())

  #wind dir
  if group['Wind Dir (10s deg)'].mean() == np.nan:
    wind_dir.append(month_weather[month_weather['Month'] == month]['Wind Dir (10s deg)'])
  else:
    wind_dir.append(group['Wind Dir (10s deg)'].mean())

  #wind speed
  if group['Wind Spd (km/h)'].mean() == np.nan:
    wind_speed.append(month_weather[month_weather['Month'] == month]['Wind Spd (km/h)'])
  else:
    wind_speed.append(group['Wind Spd (km/h)'].mean())

  #visibility
  if group['Visibility (km)'].mean() == np.nan:
    visibility.append(month_weather[month_weather['Month'] == month]['Visibility (km)'])
  else:
    visibility.append(group['Visibility (km)'].mean())

  #pressure
  if group['Stn Press (kPa)'].mean() == np.nan:
    pressure.append(month_weather[month_weather['Month'] == month]['Stn Press (kPa)'])
  else:
    pressure.append(group['Stn Press (kPa)'].mean())


df = df_weather.groupby(['Year', 'Month', 'Day', 'period']).sum().reset_index()
df['Weather'] = weather
df['Temp (°C)'] = temp
df['Dew Point Temp (°C)'] = dew_temp
df['Rel Hum (%)'] = rh
df['Wind Dir (10s deg)'] = wind_dir
df['Wind Spd (km/h)'] = wind_speed
df['Visibility (km)'] = visibility
df['Stn Press (kPa)'] = pressure

df = df.drop(columns = ['index', 'Longitude (x)','Latitude (y)','hour_of_day'])

df.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_weather.csv')

df

#Bike Data

year = []
month = []
day = []

for time in df_bike['start_time']:
  year.append(time.year)
  month.append(time.month)
  day.append(time.day)

df_bike['Year'] = year
df_bike['Month'] = month
df_bike['Day'] = day

df_bike = df_bike.drop(df_bike[df_bike['Year']==2016].index)

df_bike['user_type'].unique()

df_bike['user_type'] = df_bike['user_type'].replace('Annual Member', 'Member')
df_bike['user_type'] = df_bike['user_type'].replace('Casual Member', 'Casual')

df_bike['user_type'].unique()

df_bike.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/bike_data.csv')

df_bike.head()

num_trips = []
num_member = []
num_casual = []
mean_duration = []

for index, group in df_bike.groupby(['Year', 'Month', 'Day', 'period']):
  num_trips.append(group['trip_id'].count())
  num_member.append(group[group['user_type']=='Member'].count()['trip_id'])
  num_casual.append(group[group['user_type']=='Casual'].count()['trip_id'])
  mean_duration.append(group['duration'].mean())

df_bike_period = df_bike.groupby(['Year', 'Month', 'Day', 'period']).mean()
df_bike_period

df_bike_period['total trips'] = num_trips
df_bike_period['total member'] = num_member
df_bike_period['total casual'] = num_casual
df_bike_period['mean duration'] = mean_duration
df_bike_period = df_bike_period.reset_index()

df_bike_period = df_bike_period.drop(columns = ['trip_id', 'duration', 'hour_of_day', 'day_of_week'])

df_bike_period

df_bike_period['season'] = df_bike_period['season'].astype(int)

df_bike_period = df_bike_period.drop(columns = ['season'])

df_bike_period.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_bike.csv')
