==> Data_Cleaning/Step_4_weather_and_trip_2.ipynb <==
import os
import pandas as pd
import seaborn as sns
import matplotlib.pylab as plt
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

df_bike = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_bike.csv')
df_bike = df_bike.drop(columns = ['Unnamed: 0'])
df_bike

#clean the weather

df_weather = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_weather.csv')
df_weather = df_weather.drop(columns = ['Unnamed: 0'])
df_weather = df_weather.drop(df_weather[df_weather['Year']==2021][df_weather[df_weather['Year']==2021]['Month']==12].index)
df_weather

df_weather['Temp (°C)'] = df_weather['Temp (°C)'].fillna(18.020996)
df_weather['Dew Point Temp (°C)'] = df_weather['Dew Point Temp (°C)'].fillna(12.730359)
df_weather['Rel Hum (%)'] = df_weather['Rel Hum (%)'].fillna(72.455093)
df_weather['Wind Dir (10s deg)'] = df_weather['Wind Dir (10s deg)'].fillna(19.89462920633175)
df_weather['Wind Spd (km/h)'] = df_weather['Wind Spd (km/h)'].fillna(14.42541)
df_weather['Visibility (km)'] = df_weather['Visibility (km)'].fillna(15.119552)
df_weather['Stn Press (kPa)'] = df_weather['Stn Press (kPa)'].fillna(100.421307)

df_weather.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_weather_2.csv')

df_weather

#concat trip and weather

header = ['Year__', 'Month__', 'Day__', 'period__', 'total trips', 'total member',
       'total casual', 'mean duration']
df_bike.columns = header

df = pd.concat([df_weather, df_bike], axis = 1)

df = df.drop(columns = ['Year__', 'Month__', 'Day__', 'period__'])

df.head()

df.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_bike_and_weather.csv')

#add true false weekend

pip install holidays

df = pd.read_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_bike_and_weather.csv')

import datetime
from datetime import date                                          
import holidays

time = []
weekday = []
ca_holidays = holidays.Canada()    
holiday_lst = [] 
weekend_lst = []                    


for idx, row in df.iterrows():
  date = datetime.datetime(row['Year'], row['Month'], row['Day'])
  time.append(date)
  weekday.append(date.weekday())

  #check whether the date is holiday
  holiday_name = ca_holidays.get(date)
  if holiday_name != None: 
    holiday_lst.append(True)
  else:
    holiday_lst.append(False)


df['Time'] = time
df['Weekday'] = weekday
df['is_holiday'] = holiday_lst

is_weekend = (df['Weekday'] == 5) | (df['Weekday'] == 6)
df['is_weekend'] = is_weekend

df.columns

header = ['Time', 'Year', 'Month', 'Day', 'period', 'Weekday', 'is_holiday', 'is_weekend', 'Temp (°C)',
       'Dew Point Temp (°C)', 'Rel Hum (%)', 'Wind Dir (10s deg)',
       'Wind Spd (km/h)', 'Visibility (km)', 'Stn Press (kPa)', 'Weather',
       'total trips', 'total member', 'total casual', 'mean duration']
df = df[header]
df

df = df.rename(columns={'Temp (°C)': "Mean Temp (°C)",
                   "period": "Period",
                   "total trips": "Total Trips",
                   "total member": "Member Trips",
                   "total casual": "Casual Trips",
                   "mean duration": "Mean Trip Duration (sec)",
                   "A": "a",})
df.head(50)

df.to_csv('/content/drive/MyDrive/CIV1498 Final Project/Data/cleaned_data3/period_bike_and_weather_new_header.csv')

df.head(50)

df
